feat(football): Added JSON support for `get_football_data` handler.
This commit adds full E2E JSON support for the football data handler. Where before the outgoing request was serialising the JSON response as a String, the response is now considered to be a full-fledged JSON object both within the API code and in the response it sends out.

This also has the added benefit of axum automatically adding in the `content-type: application/json` header for us. Neat!