feat(football_results): Removed strict JSON tyoe safety as we're operating as a pass-through layer in this API.
I've elected to simply treat the JSON as a shapeless serde_json::Value here. In addition to simplifying this web server, this means changes beyond my control to the backing API will not cause breakages at this layer, as well as ensuring consistency between the frontend and the backing API, converting this layer to a "real" proxy.